edit , add and isbn type fixed

+++ src/main/java/com/adminportal/domain/Book.java
@@ -23,7 +23,7 @@ public class Book {
     private String category;
     private int numberOfPages;
     private String format;
-    private int isbn;
+    private long isbn;
     private double shippingWeight;
     private double listPrice;
     private double ourPrice;
@@ -108,11 +108,11 @@ public class Book {
         this.format = format;
     }
 
-    public int getIsbn() {
+    public long getIsbn() {
         return isbn;
     }
 
-    public void setIsbn(int isbn) {
+    public void setIsbn(long isbn) {
         this.isbn = isbn;
     }
